# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gatewayservice API
  description: Gateway OpenAPI specification.
  version: 0.2.0
servers:
  - url: http://localhost:8000
    description: Gateway service (local)
  - url: http://79.72.56.202:8000
    description: Gateway service (remote)
...
paths:
...
  /question/{questionType}:
    get:
      summary: Fetch a question from the database.
      operationId: fetchQuestion
      tags:
        - Questions
      parameters:
        - name: questionType
          in: path
          required: true
          description: The type of the question to retrieve
          schema:
            type: string
            enum: ["flag", "car", "famous-person", "dino", "place"]
            example: "flag"
      responses:
        "200":
          description: The system returns a question.
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    description: Unique identifier for the question.
                  type:
                    type: string
                    description: Type of question.
                    example: "flag"
                  imageUrl:
                    type: string
                    format: uri
                    description: URL of an image related to the question.
                    example: "http://commons.wikimedia.org/wiki/Special:FilePath/Flag%20of%20Spain.svg"
                  options:
                    type: array
                    description: List of possible answers.
                    items:
                      type: string
                    example:
                      - "Netherlands"
                      - "France"
                      - "Syria"
                      - "Egypt"
                  correctAnswer:
                    type: string
                    description: The correct answer.
                    example: "Spain"
                  alreadyShown:
                    type: boolean
                    description: Indicates if the question has been shown before.
                    example: true
                  createdAt:
                    type: string
                    format: date-time
                    description: Timestamp when the question was created.
                    example: "2025-03-18T18:37:03.978Z"
        "404":
          description: No questions available.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: No questions available.
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /fetch-question-data:
    post:
      summary: Fetches questions from Wikidata and stores them in the database.
      operationId: fetchFlagData
      tags:
        - Questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                numberOfQuestions:
                  type: integer
                  description: Number of questions to fetch.
                  example: 10
                questionType:
                  oneOf:
                    - type: string
                      enum:
                        - flag
                        - car
                        - famous-person
                        - dino
                        - place
                  description: The model to use for the LLM.
      responses:
        "200":
          description: Questions fetched and stored successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                      description: The type of quiz question.
                      example: "flag"
                    imageUrl:
                      type: string
                      format: uri
                      description: URL of an image related to the question.
                      example: "http://commons.wikimedia.org/wiki/Special:FilePath/Flag%20of%20Japan.svg"
                    options:
                      type: array
                      description: List of possible answers.
                      items:
                        type: string
                      example:
                        - "Japan"
                        - "India"
                        - "Korea"
                        - "China"
                    correctAnswer:
                      type: string
                      description: The correct answer to the question.
                      example: "Japan"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
...
  /check-answer:
    post:
      summary: Check the answer to a question.
      operationId: checkAnswer
      tags:
        - Questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                questionId:
                  type: string
                  description: The ID of the question.
                  example: "questionId1"
                selectedAnswer:
                  type: string
                  description: The user's answer.
                  example: "Spain"
      responses:
        "200":
          description: Answer checked successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  isCorrect:
                    type: boolean
                    description: Indicates if the answer is correct.
                    example: true
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Invalid question ID or answer.
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
...